cleared some general stuff

+++ src/app/components/main-nav/main-nav.component.ts
@@ -1,6 +1,7 @@
 import { Component, inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { UserService } from 'src/app/services/user.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'main-nav',
@@ -12,9 +13,35 @@ export class MainNavComponent {
   router = inject(Router)
   userService = inject(UserService)
 
-  onLogout() {
-    localStorage.clear()
-    this.router.navigate(['/home'])
+  async onLogout() {
+    const result = await Swal.fire({
+      title: "Do you want to log out?",
+      text: "Your local storage will be cleared!",
+      showCancelButton: true,
+      confirmButtonText: "Yes",
+      cancelButtonText: "No",
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+    })
+    if (result.isConfirmed) {
+      localStorage.clear()
+      this.router.navigate(['/home'])
+    }
   }
 
+
+  // async onClickLogout() {
+  //   const result = await Swal.fire({
+  //     title: "Do you want to log out?",
+  //     icon: "warning",
+  //     showCancelButton: true,
+  //     confirmButtonColor: "#3085d6",
+  //     cancelButtonColor: "#d33",
+  //     confirmButtonText: "Yes",
+  //     cancelButtonText: "No",
+  //   })
+  //   if (result.isConfirmed) {
+  //     localStorage.removeItem('auth_token')
+  //     this.router.navigate(['/login'])
+  //   }
 }
